refactor: make CLI implementation modular
