+++ src/cli/__init__.py
@@ -1,0 +1,7 @@
+"""
+Ambitus AI Models CLI Package
+
+A comprehensive CLI tool for running AI-powered market research agents.
+"""
+
+__version__ = "0.1.0"
