+++ src/cli/commands/__init__.py
@@ -1,0 +1,1 @@
+"""CLI commands package"""
