+++ src/cli/main.py
@@ -1,0 +1,31 @@
+import click
+import sys
+from pathlib import Path
+
+# Add project root to path
+project_root = Path(__file__).parent.parent.parent
+sys.path.insert(0, str(project_root))
+
+from src.cli.commands.api import api_command
+from src.cli.commands.mcp import mcp_command
+from src.cli.commands.pipeline import pipeline_command
+from src.cli.commands.tui import tui_command
+
+@click.group()
+@click.version_option(version="0.1.0")
+def main():
+    """
+    Ambitus AI Models - Market Research Automation Platform
+    
+    A comprehensive CLI tool for running AI-powered market research agents.
+    """
+    pass
+
+# Register commands
+main.add_command(api_command)
+main.add_command(mcp_command)
+main.add_command(pipeline_command)
+main.add_command(tui_command)
+
+if __name__ == "__main__":
+    main()
