+++ src/cli/tui/__init__.py
@@ -1,0 +1,1 @@
+"""Terminal User Interface package"""
